Move removeSlashes localy to transip and use it only for TXT
