+++ providers/transip/slashes.go
@@ -1,10 +1,10 @@
-package models
+package transip
 
 import (
 	"regexp"
 )
 
-func RemoveSlashes(s string) string {
+func removeSlashes(s string) string {
 	m := regexp.MustCompile("(?:\\\\(\\\\)+)|(?:\\\\)")
 	return m.ReplaceAllString(s, "$1")
 }
